bug: changing centering vector gives different homology groups
changing the centering vector of the fundamental domain gives different homology groups. further, these groups are not torsion, so by a theorem giving the Q-co-homology should not be there. this suggests there's a bug in the boundary maps. however, given that a very complex procedure gives a complex, and any change to the procedure ruins the differential squaring to zero, it seems like the problem must lie with the cellular decomposition. specifically, some faces have non-trivial stabilizer. a finer, barycentric triangulation will indeed remove stabilizers, but then there are way too many vertices for the computation to proceed (in the simplest example, you go from 7 -> 78 vertices, which takes over a day and did not finish overnight).