added cleaner mod2 support
explicitly take modulo 2 in boundary map, though it shouldn't matter when you perform do that. it does highlight where the calculation may be going wrong - trying to correct the count for faces that are stabilized. it seems like you need to remove fixed points and get rid of cells that are stabilized to compute anything sensible.