network bubble sort seems to find the min
for small networks, #vertices < 10, network bubble sort seems to find the minimum, or very close to it. potential bug in that sometimes the minimum found by sampling all permutations is larger than bubble sort. given that the permutations used in bubble sort are a subset of all permutations, this doesn't make sense.